pyproject.toml
[tool.poetry]
name = "drf-activities"
version = "0.1.0"
description = ""
authors = ["Luciana <luciheit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.3.1"
Django = "3.1.6"
djangorestframework = "3.12.2"
pytz = "2021.1"
sqlparse = "0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
